1271 data api convert mocha to jest (#1426)
* Rename test folder

* Convert mocha to jest

* Add custom jest setup file

* Resolve jest hanging issue

* Delete .babelrc

* Rename jestScaffolding to jest.setup

* Update eslintConfig in package.json

* Update test:converage

* Remove winston.config

* Use toBeRejectedWith
