+++ packages/data-api/jest.config.json
@@ -1,0 +1,8 @@
+{
+  "clearMocks": true,
+  "collectCoverageFrom": ["<rootDir>/src/**/*.js"],
+  "setupFilesAfterEnv": ["../../jestScaffolding.js", "./jest.setup.js"],
+  "testMatch": ["<rootDir>/src/__tests__/**/**.test.js"],
+  "transform": { "^.+\\.jsx?$": "../../jestTransformer.js" },
+  "transformIgnorePatterns": ["/node_modules/"]
+}
